requirement.txt
fastapi
uvicorn
qdrant-client
httpx
pypdf
python-docx
google-generativeai
tiktoken
python-multipart
python-dotenv
fitz
pdfplumber